pyproject.toml
[tool.poetry]
name = "aoc"
version = "0.1.0"
description = ""
authors = ["Svetoslav Karaivanov <first_name.second_name@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
black = "^24.10.0"
isort = "^5.12.0"
tqdm = "^4.66.1"
beautifulsoup4 = "^4.12.2"
requests = "^2.31.0"
numpy = "^2.1.3"
pyyaml = "^6.0.1"
click = "^8.1.7"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.27.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
